+++ src/context/PokemonContext.tsx
@@ -1,11 +1,4 @@
-import {
-	createContext,
-	useCallback,
-	useContext,
-	useEffect,
-	useState,
-} from "react";
-import type { ReactNode } from "react";
+import { type ReactNode, createContext, useContext, useState } from "react";
 
 export interface PokemonData {
 	id: number;
@@ -16,25 +9,9 @@ export interface PokemonData {
 	types: string[];
 }
 
-interface NameEntry {
-	name: string;
-	language: {
-		name: string;
-		url: string;
-	};
-}
-
-interface TypeEntry {
-	slot: number;
-	type: {
-		name: string;
-		url: string;
-	};
-}
-
 interface PokemonContextType {
 	pokemonData: PokemonData[];
-	refreshPokemons: () => void;
+	setPokemonData: (data: PokemonData[]) => void;
 }
 
 const PokemonContext = createContext<PokemonContextType | undefined>(undefined);
@@ -42,60 +19,8 @@ const PokemonContext = createContext<PokemonContextType | undefined>(undefined);
 export const PokemonProvider = ({ children }: { children: ReactNode }) => {
 	const [pokemonData, setPokemonData] = useState<PokemonData[]>([]);
 
-	const getRandomPokemonIds = useCallback(() => {
-		const ids = new Set<number>();
-		while (ids.size < 30) {
-			ids.add(Math.floor(Math.random() * 251) + 1);
-		}
-		return Array.from(ids);
-	}, []);
-
-	const fetchPokemons = useCallback(async () => {
-		try {
-			const ids = getRandomPokemonIds();
-
-			const promises = ids.map(async (id) => {
-				const [pokemonRes, speciesRes] = await Promise.all([
-					fetch(`https://pokeapi.co/api/v2/pokemon/${id}`).then((res) =>
-						res.json(),
-					),
-					fetch(`https://pokeapi.co/api/v2/pokemon-species/${id}`).then((res) =>
-						res.json(),
-					),
-				]);
-
-				const frenchNameEntry = speciesRes.names.find(
-					(entry: NameEntry) => entry.language.name === "fr",
-				);
-				const frenchName = frenchNameEntry
-					? frenchNameEntry.name
-					: pokemonRes.name;
-
-				return {
-					id: pokemonRes.id,
-					name: frenchName,
-					image: pokemonRes.sprites.front_default,
-					weight: pokemonRes.weight,
-					height: pokemonRes.height,
-					types: pokemonRes.types.map((t: TypeEntry) => t.type.name),
-				};
-			});
-
-			const clearData = await Promise.all(promises);
-			setPokemonData(clearData);
-		} catch (error) {
-			console.error("Erreur de chargement des Pokémon:", error);
-		}
-	}, [getRandomPokemonIds]);
-
-	useEffect(() => {
-		fetchPokemons();
-	}, [fetchPokemons]);
-
 	return (
-		<PokemonContext.Provider
-			value={{ pokemonData, refreshPokemons: fetchPokemons }}
-		>
+		<PokemonContext.Provider value={{ pokemonData, setPokemonData }}>
 			{children}
 		</PokemonContext.Provider>
 	);
