+++ src/pages/game/Game.tsx
@@ -6,11 +6,64 @@ import TrainerCheck from "../../components/trainerCheck/TrainerCheck";
 import { usePokemonContext } from "../../context/PokemonContext";
 import type { PokemonData } from "../../context/PokemonContext";
 
+interface NameEntry {
+	name: string;
+	language: { name: string; url: string };
+}
+
+interface TypeEntry {
+	slot: number;
+	type: { name: string; url: string };
+}
+
+const getRandomPokemonIds = () => {
+	const ids = new Set<number>();
+	while (ids.size < 30) {
+		ids.add(Math.floor(Math.random() * 251) + 1);
+	}
+	return Array.from(ids);
+};
+
 function Game() {
-	const { pokemonData } = usePokemonContext();
+	const { pokemonData, setPokemonData } = usePokemonContext();
 	const [currentIndex, setCurrentIndex] = useState(0);
 	const [shuffledPokemon, setShuffledPokemon] = useState<PokemonData[]>([]);
 	const [selectedPokemons, setSelectedPokemons] = useState<PokemonData[]>([]);
+
+	useEffect(() => {
+		const fetchPokemons = async () => {
+			const ids = getRandomPokemonIds();
+
+			const promises = ids.map(async (id) => {
+				const [pokemonRes, speciesRes] = await Promise.all([
+					fetch(`https://pokeapi.co/api/v2/pokemon/${id}`).then((res) =>
+						res.json(),
+					),
+					fetch(`https://pokeapi.co/api/v2/pokemon-species/${id}`).then((res) =>
+						res.json(),
+					),
+				]);
+
+				const frenchNameEntry = speciesRes.names.find(
+					(entry: NameEntry) => entry.language.name === "fr",
+				);
+
+				return {
+					id: pokemonRes.id,
+					name: frenchNameEntry?.name || pokemonRes.name,
+					image: pokemonRes.sprites.front_default,
+					weight: pokemonRes.weight,
+					height: pokemonRes.height,
+					types: pokemonRes.types.map((t: TypeEntry) => t.type.name),
+				};
+			});
+
+			const clearData = await Promise.all(promises);
+			setPokemonData(clearData);
+		};
+
+		fetchPokemons();
+	}, [setPokemonData]);
 
 	useEffect(() => {
 		if (pokemonData.length >= 30) {
@@ -55,16 +108,12 @@ function Game() {
 						<img
 							src="src/assets/images/test_img/test_temoin.svg"
 							alt="Le témoin officiel de tous les pokedexpatrolleurs"
-							onClick={handleClick}
-							onKeyDown={(e) => e.key === "a" && handleClick()}
 						/>
 					</div>
 					<div className="pokedex">
 						<img
 							src="src/assets/images/test_img/test_pokedex.svg"
 							alt="Ceci est un pokédex"
-							onClick={handleClick}
-							onKeyDown={(e) => e.key === "a" && handleClick()}
 						/>
 					</div>
 
@@ -79,8 +128,6 @@ function Game() {
 						<img
 							src="src/assets/images/test_img/test_permistrainer.svg"
 							alt="Ceci est le permis du dresseur"
-							onClick={handleClick}
-							onKeyDown={(e) => e.key === "a" && handleClick()}
 						/>
 					</div>
 				</div>
